ai_backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

# PDF parsing
PyPDF2==3.0.1

# NLP
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# ML (CPU-only, critical for Railway)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.2+cpu
sentence-transformers==2.5.1

# Gemini
google-generativeai==0.4.1
python-dotenv==1.0.1

# Sentiment
vaderSentiment==3.3.2
